Feature/26074 reorganise orchestrator functions (#993)
* wip

* wip

* wip

+++ polaris-pipeline/coordinator/Durable/Payloads/BasePipelinePayload.cs
@@ -8,7 +8,7 @@ namespace coordinator.Durable.Payloads;
 
 public abstract class BasePipelinePayload
 {
-    protected BasePipelinePayload(string cmsCaseUrn, long cmsCaseId, Guid correlationId, PolarisDocumentId polarisDocumentId = null)
+    protected BasePipelinePayload(string cmsCaseUrn, int cmsCaseId, Guid correlationId, PolarisDocumentId polarisDocumentId = null)
     {
         CmsCaseUrn = cmsCaseUrn;
         CmsCaseId = cmsCaseId;
@@ -20,7 +20,7 @@ public abstract class BasePipelinePayload
     public string CmsCaseUrn { get; set; }
 
     [RequiredLongGreaterThanZero]
-    public long CmsCaseId { get; set; }
+    public int CmsCaseId { get; set; }
 
     [Required]
     public Guid CorrelationId { get; set; }
