+++ polaris-pipeline/coordinator/Durable/Payloads/GetCaseDocumentsActivityPayload.cs
@@ -4,7 +4,7 @@ namespace coordinator.Durable.Payloads
 {
     public class GetCaseDocumentsActivityPayload : BasePipelinePayload
     {
-        public GetCaseDocumentsActivityPayload(string caseUrn, long caseId, string cmsAuthValues, Guid correlationId) :
+        public GetCaseDocumentsActivityPayload(string caseUrn, int caseId, string cmsAuthValues, Guid correlationId) :
             base(caseUrn, caseId, correlationId)
         {
             CmsAuthValues = cmsAuthValues;
